+++ bupocket/Classes/Assets/Model/RegisteredModel.h
@@ -14,7 +14,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property (nonatomic, copy) NSString * name;
 @property (nonatomic, copy) NSString * code;
-@property (nonatomic, assign) NSInteger amount;
+@property (nonatomic, assign) int64_t amount;
 @property (nonatomic, assign) NSInteger decimals;
 @property (nonatomic, copy) NSString * desc;
